Refactor VertexAI
