+++ core/llm/llms/index.ts
@@ -8,7 +8,6 @@ import {
 import { renderTemplatedString } from "../../promptFiles/renderTemplatedString";
 import { BaseLLM } from "../index";
 import Anthropic from "./Anthropic";
-import AnthropicVertexAI from "./AnthropicVertexAI";
 import Azure from "./Azure";
 import Bedrock from "./Bedrock";
 import BedrockImport from "./BedrockImport";
@@ -21,7 +20,6 @@ import Fireworks from "./Fireworks";
 import Flowise from "./Flowise";
 import FreeTrial from "./FreeTrial";
 import Gemini from "./Gemini";
-import GeminiVertexAI from "./GeminiVertexAI";
 import Groq from "./Groq";
 import HuggingFaceInferenceAPI from "./HuggingFaceInferenceAPI";
 import HuggingFaceTGI from "./HuggingFaceTGI";
@@ -30,7 +28,6 @@ import LMStudio from "./LMStudio";
 import LlamaCpp from "./LlamaCpp";
 import Llamafile from "./Llamafile";
 import Mistral from "./Mistral";
-import MistralVertexAI from "./MistralVertexAI";
 import Mock from "./Mock";
 import Msty from "./Msty";
 import Nvidia from "./Nvidia";
@@ -45,14 +42,13 @@ import Together from "./Together";
 import Vllm from "./Vllm";
 import WatsonX from "./WatsonX";
 import ContinueProxy from "./stubs/ContinueProxy";
+import VertexAI from "./VertexAI";
 
 const LLMs = [
   Anthropic,
-  AnthropicVertexAI,
   Cohere,
   FreeTrial,
   Gemini,
-  GeminiVertexAI,
   Llamafile,
   Ollama,
   Replicate,
@@ -65,7 +61,6 @@ const LLMs = [
   OpenAI,
   LMStudio,
   Mistral,
-  MistralVertexAI,
   Bedrock,
   BedrockImport,
   SageMaker,
@@ -85,6 +80,7 @@ const LLMs = [
   SambaNova,
   Mock,
   Cerebras,
+  VertexAI
 ];
 
 export async function llmFromDescription(
