+++ docs/docs/customize/model-providers/top-level/vertexai.md
@@ -1,6 +1,6 @@
 ---
-title: Mistral Vertex AI
-slug: ../mistral-vertexai
+title: Vertex AI
+slug: ../vertexai
 ---
 
 :::info
@@ -11,14 +11,14 @@ You need to enable the [Vertex AI API](https://console.cloud.google.com/marketpl
 
 ## Chat model
 
-We recommend configuring **Mistral Large** as your chat model.
+We recommend configuring **Mistral Large**, **Gemini 1.5 Pro**, or **Claude 3.5 Sonnet** as your chat model.
 
 ```json title="config.json"
 {
   "models": [
     {
       "title": "Mistral Large (Vertex AI)",
-      "provider": "mistral-vertexai",
+      "provider": "vertexai",
       "model": "mistral-large",
       "projectId": "[PROJECT_ID]",
       "region": "us-central1"
@@ -27,15 +27,56 @@ We recommend configuring **Mistral Large** as your chat model.
 }
 ```
 
+```json title="config.json"
+{
+  "models": [
+    {
+      "title": "Gemini 1.5 Pro (Vertex AI)",
+      "provider": "vertexai",
+      "model": "gemini-1.5-pro",
+      "projectId": "[PROJECT_ID]",
+      "region": "us-central1"
+    }
+  ]
+}
+```
+
+```json title="config.json"
+{
+  "models": [
+    {
+      "title": "Claude 3.5 Sonnet",
+      "provider": "vertexai",
+      "model": "claude-3-5-sonnet-20240620",
+      "projectId": "[PROJECT_ID]",
+      "region": "us-east5"
+    }
+  ]
+}
+```
+
+
 ## Autocomplete model
 
-We recommend configuring **Codestral** as your autocomplete model.
+We recommend configuring **Codestral** or **code-gecko** as your autocomplete model.
 
 ```json title="config.json"
 {
   "tabAutocompleteModel": {
     "title": "Codestral (Vertex AI)",
-    "provider": "mistral-vertexai",
+    "provider": "vertexai",
+    "model": "codestral",
+    "projectId": "[PROJECT_ID]",
+    "region": "us-central1"
+  }
+}
+```
+
+```json title="config.json"
+{
+  "tabAutocompleteModel": {
+    "title": "Code Gecko (Vertex AI)",
+    "provider": "vertexai",
     "model": "codestral",
     "projectId": "[PROJECT_ID]",
     "region": "us-central1"
@@ -50,7 +91,7 @@ We recommend configuring **text-embedding-004** as your embeddings model.
 ```json title="config.json"
 {
   "embeddingsProvider": {
-    "provider": "vertex",
+    "provider": "vertexai",
     "model": "text-embedding-004",
     "projectId": "[PROJECT_ID]",
     "region": "us-central1"
@@ -60,6 +101,6 @@ We recommend configuring **text-embedding-004** as your embeddings model.
 
 ## Reranking model
 
-Mistral Vertex AI currently does not offer any reranking models.
+Vertex AI currently does not offer any reranking models.
 
 [Click here](../../model-types/reranking.md) to see a list of reranking model providers.
