+++ gui/src/pages/AddNewModel/configs/models.ts
@@ -791,6 +791,33 @@ export const models: { [key: string]: ModelPackage } = {
     providerOptions: ["watsonx"],
     isOpenSource: false,
   },
+  VertexGemini15Pro: {
+    title: "Gemini 1.5 Pro",
+    description: "A newer Gemini model with 1M token context length",
+    params: {
+      title: "Gemini 1.5 Pro",
+      model: "gemini-1.5-pro-002",
+      contextLength: 2_097_152,
+    },
+    icon: "gemini.png",
+    providerOptions: ["vertexai"],
+    isOpenSource: false,
+  },
+  VertexGemini15Flash: {
+    title: "Gemini 1.5 Flash",
+    description:
+      "Fast and versatile multimodal model for scaling across diverse tasks",
+    params: {
+      title: "Gemini 1.5 Flash",
+      model: "gemini-1.5-flash-002",
+      contextLength: 1_048_576,
+    },
+    icon: "gemini.png",
+    providerOptions: ["vertexai"],
+    isOpenSource: false,
+  },
+
+
   AUTODETECT: {
     title: "Autodetect",
     description:
